+++ test/app.test.ts
@@ -7,7 +7,7 @@ import * as config from '../src/config';
 import * as type from '../src/common/types';
 import { logger, logErr } from '../src/utils/logger';
 import * as mongo from '../src/mongo/mongo';
-import * as filters from '../src/utils/nunjuck-custom-filters';
+import * as filters from '../src/utils/nunjucks-custom-filters';
 import  app  from '../src/app';
 
 jest.mock('../src/mongo/mongo');
